Fix flakey DB test
